Reduces Terraform churn for NICs
TL;DR
-----

Fills in details on NICs to prevent Terraform churn

Details
-------

Terraform was showing differences in the plan for the VMKernel
NICSs because there the code used the defaults for IPv6. This
code makes the defaults explicit so that the Terraform plan will
not see a change that isn't.